requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2
requests>=2.31.0 

# AI & ML
anthropic>=0.18.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0

# Background tasks (choose one based on your setup)
# celery>=5.3.0  # For Celery
# django-q>=1.3.9  # For Django-Q
# OR use Django's built-in async for simple cases

# Database
# psycopg2-binary>=2.9.0  # For PostgreSQL (recommended for production)
# OR
# mysqlclient>=2.2.0  # For MySQL

# Utilities
# python-dotenv>=1.0.0  # For environment variables